backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
openai==1.3.6
requests==2.31.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.2
python-dateutil==2.8.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
cryptography==41.0.7
PyJWT==2.8.1